pyproject.toml
[tool.poetry]
name = "tcap-alb-logs-ingestion"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tcap_alb_logs_ingestion"}]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.26.78"
pandas = "^1.5.3"
configparser = "^5.3.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
